+++ src/data/combination_db.lua
@@ -1,0 +1,4 @@
+local Database = require 'util.database'
+local COMBINATIONS = Database(7)
+
+return COMBINATIONS