+++ src/data/combinations.lua
@@ -1,7 +1,0 @@
--- 填入合成材料
-
-COMBINATIONS = {
-    {"sbch", "crys"}
-}
-
-return COMBINATIONS